Discord Bot Python File
This is a Python file that I wrote for a Discord bot that accepts user commands and also performs different server related functions. This is the only file that you will need in order to get a bot running on your server, all you have to do is add your bot token.